+++ tests/run_tests.sh
@@ -6,7 +6,7 @@ ROOT_DIR=$(realpath $(dirname "$0")/..)
 
 build() {
     FLAVOR="${1}"
-    nvidia-docker build \
+    docker build \
         --pull \
         --build-arg UNAME=$(whoami) \
         --build-arg UID=$(id -u) \
@@ -19,18 +19,23 @@ build() {
 exec_cpu() {
     FLAVOR="${1}"
     shift
-    cd ${ROOT_DIR} && nvidia-docker run \
-        -it --rm --network=host \
+    cd ${ROOT_DIR} && docker run \
+        -it --rm --network=host --ipc=host \
         -v "${ROOT_DIR}":/work \
         "torch-fidelity-test-${FLAVOR}" \
         $@
 }
 
 exec_cuda() {
+    if [ -z "${CUDA_VISIBLE_DEVICES}" ]; then
+        echo "CUDA_VISIBLE_DEVICES not set, using CUDA_VISIBLE_DEVICES=0"
+        export CUDA_VISIBLE_DEVICES=0
+    fi
     FLAVOR="${1}"
     shift
     cd ${ROOT_DIR} && nvidia-docker run \
-        -it --rm --network=host \
+        --gpus "device=${CUDA_VISIBLE_DEVICES}" \
+        -it --rm --network=host --ipc=host \
         --env CUDA_VISIBLE_DEVICES=0 \
         -v "${ROOT_DIR}":/work \
         "torch-fidelity-test-${FLAVOR}" \
@@ -40,14 +45,14 @@ exec_cuda() {
 main() {
     FLAVOR="${1}"
     build "${FLAVOR}"
-    exec_cpu "${FLAVOR}" python3 -m unittest discover "tests/${FLAVOR}"
+    exec_cuda "${FLAVOR}" python3 -W error -m unittest discover "tests/${FLAVOR}"
 }
 
 main_sh() {
     FLAVOR="${1}"
     build "${FLAVOR}"
     for test in tests/${FLAVOR}/test_*.sh; do
-        exec_cpu "${FLAVOR}" sh ${test}
+        exec_cuda "${FLAVOR}" sh ${test}
     done
 }
 
@@ -57,10 +62,10 @@ shell() {
     exec_cuda "${FLAVOR}" bash
 }
 
-main tf1 || true
-main torch_pure || true
-main clip || true
-main backend || true
-main prc_ppl_reference || true
+main tf1
+main torch_pure
+main clip
+main prc_ppl_reference
+main_sh sphinx_doc
 
-main_sh sphinx_doc || true
+echo "=== TESTS FINISHED ==="