+++ tests/tf1/reference/reference_metric_fid_ttur.py
@@ -17,7 +17,7 @@ import numpy as np
 import tensorflow as tf
 from imageio import imread
 from scipy import linalg
-import fwr13y.d9m.tensorflow as tf_determinism
+from tfdeterminism import patch as patch_tensorflow_for_determinism
 from tqdm import tqdm
 
 # InceptionV3 pretrained weights from TensorFlow models library
@@ -308,7 +308,7 @@ if __name__ == "__main__":
 
     if args.determinism:
         with redirect_stdout(sys.stderr):
-            tf_determinism.enable_determinism()
+            patch_tensorflow_for_determinism()
 
     metrics = calculate_fid_given_paths(args.path, low_profile=args.lowprofile, verbose=not args.silent)