+++ tests/tf1/reference/reference_metric_kid_mmdgan.py
@@ -17,7 +17,7 @@ import tensorflow as tf
 from PIL import Image
 from scipy import linalg
 from sklearn.metrics.pairwise import polynomial_kernel
-import fwr13y.d9m.tensorflow as tf_determinism
+from tfdeterminism import patch as patch_tensorflow_for_determinism
 from tqdm import tqdm
 
 # InceptionV3 pretrained weights from TensorFlow models library
@@ -423,7 +423,7 @@ def main():
 
     if args.determinism:
         with redirect_stdout(sys.stderr):
-            tf_determinism.enable_determinism()
+            patch_tensorflow_for_determinism()
 
     if (args.do_fid or args.do_mmd) and args.reference_feats is None:
         parser.error("Need REFERENCE_FEATS if you're doing FID or MMD")